Added XML file as external data source
Data source is xml file for GET and POST operations in Rest Service

+++ MLSListingsRestService/Models/HomeDetails.cs
@@ -11,6 +11,6 @@ namespace MLSListingsRestService.Models
         public string Address { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public int Zip { get; set; }
+        public long Zip { get; set; }
     }
 }
